% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{stability}
\alias{stability}
\title{Check Stability of DSGE Model}
\usage{
stability(x)
}
\arguments{
\item{x}{A \code{dsge_fit} or \code{dsge_solution} object.}
}
\value{
A list of class \code{"dsge_stability"} with:
\describe{
\item{stable}{Logical: is the system saddle-path stable?}
\item{eigenvalues}{Complex eigenvalue vector.}
\item{moduli}{Moduli of eigenvalues.}
\item{classification}{Character vector: "stable" or "unstable" for each.}
\item{n_stable}{Number of stable eigenvalues.}
\item{n_states}{Number of state variables (required stable count).}
}
}
\description{
Checks saddle-path stability of the model by examining eigenvalues.
A model is stable when the number of eigenvalues with modulus less
than 1 equals the number of state variables.
}
